git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@16218 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/vips.c
@@ -645,8 +645,8 @@ static MagickBooleanType WriteVIPSImage(const ImageInfo *image_info,
     (void) WriteBlobLSBLong(image,VIPS_MAGIC_LSB);
   else
     (void) WriteBlobLSBLong(image,VIPS_MAGIC_MSB);
-  (void) WriteBlobLong(image,image->columns);
-  (void) WriteBlobLong(image,image->rows);
+  (void) WriteBlobLong(image,(unsigned int) image->columns);
+  (void) WriteBlobLong(image,(unsigned int) image->rows);
   (void) SetImageStorageClass(image,DirectClass,exception);
   channels=image->alpha_trait == BlendPixelTrait ? 4 : 3;
   if (IsImageGray(image,exception) != MagickFalse)
